Remove walk dependency and use concurrent workers
Removed the dependency on walk, which performed a parallel walk through a
directory structure, and replaced it with `filepath/walk` from stdlib, and a
pool of workers executing callbacks in parallel, as it goes, coordinated by a
channel.

This has actually made it slower in one of my tests on a 68000 file directory
structure (3.2 sec vs 1.3 sec on my MBP), probably because a single threaded
walk isn't as efficient. In real world usage this difference probably isn't
so prounced, because most of the effort is in whatever the callbacks are
doing, not the crawl itself.